backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.10.4
pydantic-settings==2.7.0
alembic==1.14.0
